requirements.txt
fastapi
uvicorn
requests
pymupdf
faiss-cpu
numpy
python-dotenv
openai
pydantic
